[REFACTOR] Split Google linking into its own submenu
Mirror the LinkWallet architecture for Google: extract the linking
ceremony into a dedicated LinkGoogle component routed via a new
linkAccountGoogle submenu. LinkedAccounts becomes a thin dashboard
with two Link buttons, dropping the inline ceremony state
(googleLinkMode, walletReauth, dispatch effect, linkingSocial
selectors). The DUAL_LOGIN gate now lives inside LinkGoogle as a
"coming soon" message rather than hiding the Linked Accounts entry
entirely.
