+++ src/features/island/hud/components/settings-menu/GameOptions.tsx
@@ -40,6 +40,7 @@ import { PlazaSettings } from "./plaza-settings/PlazaSettingsModal";
 import { DeveloperOptions } from "./developer-options/DeveloperOptions";
 import { LinkedAccounts } from "./linked-accounts/LinkedAccounts";
 import { LinkWallet } from "features/wallet/components/LinkWallet";
+import { LinkGoogle } from "features/auth/components/LinkGoogle";
 import { Discord } from "./general-settings/DiscordModal";
 import { DepositWrapper } from "features/goblins/bank/components/DepositGameItems";
 import { useSound } from "lib/utils/hooks/useSound";
@@ -284,6 +285,7 @@ export type SettingMenuId =
   | "blockchain"
   | "linkedAccounts"
   | "linkAccountWallet"
+  | "linkAccountGoogle"
   | "plaza"
   | "experiments"
   | "economyEditor"
@@ -365,6 +367,11 @@ export const settingMenus: Record<SettingMenuId, SettingMenu> = {
     title: translate("linkedAccounts.linkWallet"),
     parent: "linkedAccounts",
     content: LinkWallet,
+  },
+  linkAccountGoogle: {
+    title: translate("linkedAccounts.linkGoogle"),
+    parent: "linkedAccounts",
+    content: LinkGoogle,
   },
   plaza: {
     title: translate("gameOptions.plazaSettings"),
